fix: 修复不能正常收藏文件
原因是第一次启动后没有添加收藏相册，没有添加的原因是判断条件是没有检测到默认导入路径，而实际上收藏是没有导入路径的
解决方法是在对应位置添加额外的判断代码

Log: 修复不能正常收藏文件
Bug: https://pms.uniontech.com/zentao/bug-view-111917.html

+++ src/album/dbmanager/dbmanager.cpp
@@ -1235,8 +1235,8 @@ void DBManager::checkDatabase()
 
 void DBManager::insertSpUID(const QString &albumName, AlbumDBType astype, SpUID UID)
 {
-    //0.路径不存在
-    if (!defaultNotifyPathExists(UID)) {
+    //0.路径不存在，BUG#111917，只检查默认路径，不要把最近收藏也搞进来了，否则会导致后续收藏失败
+    if (UID != u_Favorite && !defaultNotifyPathExists(UID)) {
         //路径不存在则删除已有的相册
         if (!m_query->exec(QString("DELETE FROM AlbumTable3 WHERE UID=%1").arg(UID))) {
         }
